Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Thanks! Now I just have to remember to look at the README:)

Yeah, I know that's always the issue. Any ideas for how to surface this information when a user needs it?

One idea is to clean up the old out-of-date documentation so that it's easier to find the important stuff.

Another idea is to point to the README in the catapult/telemetry/telemetry/testing doc as well.

Neither of those two ideas sounds particularly good.

Patch-set: 7
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>