Add APIs to cros interface and cros platform backend class.
Add a cros interface API to start a command on the CrOS device and return the
process handle. This API will be used by the interval profiler to start the CPU
profiler during story runs.

Add cros platform backend APIs to expose the cros interface APIs to start a
command on the CrOS device, get a file from the CrOS device, push contents to a
file on the CrOS device.

Change-Id: I94fe4e625cbb57049382c845f26a401bcea0fa6e