DO NOT LAND !!. This is part of a prototype
This is a prototype for changes to catapult. The purpose of making these changes
is so that we can allow gpu tests to use the new test expectations format. The high
level changes that this CL makes are the following.

1, Make changes to Typ so that Typ can identify Flaky tests through the Flaky
tag in the expectations file. And when Typ identifies a Flaky test it will
do Flaky retries on that test. Flaky retry count can be specified through
a command line argument --retry-limit-for-flaky-tests, or it can be passed to
Typ by the browser test class by the GetFlakyRetryCount() method.

2, Allow browser tests to pass expectations files for tests methods that they have
to Typ by overriding a function, ExpectationsFiles, and
having it return a list of test expectations file paths.

3, Allow expectations file to only have the test method names instead of full
test method paths in each expectation.

R=dpranke@chromium.org, kbr@chromium.org

Bug: chromium:698902
Change-Id: Ib44ebb6165e5a652a6883c8ca0d92147293eadc7