Fixes issue with tests skipped via command line arguments
This issue happens when an expectations file is passed to Typ.
A test can be skipped by passing a command line argument, i.e
--skip=*test_pass. The test is skipped, however the expected field
shows only 'SKIP' irregardless of the actual expectations specified
in the expectations file. This CL will make a change so when a test
is skipped via a command line argument, the expectations specified
in the expectations file will show up in the expected field of the
test's JSON results.

Example:
  Test method:
  class PassTest(unittest.TestCase):
    def test_pass(self):
      return

  expectations file:
  # tags: [foo]
  crbug.com/123 [foo] PassTest.test_pass [Failure Crash]

  command:
  ./telemetry/run_tests  PassTest.test_pass --skip=*test_pass

  json results
  {
    ...
    test_pass: {
      'expected': 'SKIP' <-- should be 'CRASH FAIL'
      'actual': 'SKIP'
    }
    ...
  }

R=kbr@chromium.org

Bug: chromium:923494
Change-Id: I09725c79f73f54b2e8c0b1c380a6dede7e42fa66