Fixes issue for tests that are skipped via calling self.skipTest()
The issue happens with unit tests and when no expectations file is
passed. If a unit test calls self.skipTest(), the expected field will show
PASS instead of SKIP. We want skipped unit tests results to be consistent
with the test results of browser tests. If a browser test calls
self.skipTest() and no expectations files are passed, then the expected
field in that tests result will show SKIP.

example:

  Test method:
  class SkipTest(unittest.TestCase):
    def test_skip(self):
      self.skipTest()

  command:
  ./telemetry/run_tests  SkipTest.test_skip

  json results:
  {
    ...
    test_skip: {
      'expected' : 'PASS' <-- Should be SKIP
      'actual' : 'SKIP'
    }
    ...
  }

R=kbr@chromium.org

Bug: chromium:923494
Change-Id: I655d3064da57681445a697e78827ebfe9741e9c8