Switch pinpoint to chops-signin.
chops-signin is a new vanilla web component to replace google-signin.
It is simpler to use and more consistent with MD and other ChOps apps.
This CL copies it into pinpoint/elements for now because the chopsui
bower package depends on Polymer2, and pinpoint isn't ready to upgrade
from Polymer 1 yet.

Demo:
https://dev-benjhayden-892d5b01-dot-pinpoint-dot-chromeperf.appspot.com/

Change-Id: Id6fd53293f40af530e443962f929aa3977e4018d