Add RecentBugsRequest for v2spa.
The V1 UI allows the backend to suggest recent bugs for sheriffs to assign new
alerts. It fetches the list of recent bugs every time the sheriff triages an
alert group.
The V2 UI fetches the list of recent bugs once on load, and filters the list in
the frontend.

Bug: chromium:918193
Change-Id: Ic59721e7cc56cdbb5be8fcab36b8a2a1b8e7310a