Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> +estaab
> 
> Looking at this more (and re-reading the "talking about different kinds of flakiness and test results and retries" thread to refresh my memory ...).
> 
> I think we need to distinguish between the different senses of flakiness better. To recap, there's the existing ("blink") sense of flakiness, which is a test that is expected to produce one >1 actual values, any of which would be treated as an expected result, and there's the new "gpu" sense of flakiness, which is that we expect the test to pass but it may need to be retried. In that sense, I'd expect the Result.expected value to just be [ Pass ].
> 
> "Flaky" in the GPU sense is really more like "Slow" in the Blink sense: it changes how we run the test. "Slow" doesn't currently exist in TYP, though, so we can't unfortunately look to see how we handle that for any guidance.
> 
> I'm not quite sure what the best way to modify the code is to better capture my concerns. It may be that json_results.py shouldn't be modified at all. Or, it may be that we need to modify it, but we shouldn't use "Flaky" in the new code, for fear of confusing that with the existing senses of flaky (like the self.flaky field). Maybe we need to use one of the alternate names we thought of, like "RetryOnFailure".
> 
> Similarly, I'm reluctant to add a new --retry-limit-for-flaky-tests flag, because I think you then need to worry about the interaction between that and --retry-limit. Maybe it's better to have a --retry-only-retry-on-failure-tests flag, and then the semantics for how to handle the cases becomes clearer.
> 
> This is all kind of hand-wavy, I'm not actually sure what the right thing to do here is. To be clear: I'm not trying to change my mind on what I want to support again, I'm just trying to keep the implementation clean. Happy to talk through in person if that helps.

One of the biggest challenges of moving GPU test onto the new expectations format is providing a way to mark tests as flaky and then having only those tests be retried. Currently the way GPU tests handles flakiness is that it reruns tests marked flaky until it passes or up to N times. It is less tolerant of flakiness because it only does this for tests marked flaky. Typ is more tolerant of flakiness because it treats all tests failures as flaky. In its current implementation, Typ reruns all test failures (including expected failures) until it passes or up until N retries, where N is provided by the command line argument --retry-limit.  In my prototype I am proposing to create a new tag that can be put in the result tags for a test expectation. And when a test with that tag in its expectations fails it will be retried by Typ up to N times or until its first pass. I can remove --retry-for-flaky-tests, and reuse --retry-limit in conjunction with another flag like what you suggested, --retry-only-retry-on-failure-tests in order to only retry tests that have the new tag in its expectations. We can also decide on a tag that best describes the flakiness of GPU tests and replace Flaky with that.

Patch-set: 7
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>