Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> (2 comments)
> 
> LGTM!
> 
> I am still a bit worried that the behavior in this changelist may surprise people who read the comments and the names of things. For example the file is named "add_reserved_diagnostics", but it really has a lot more responsibilities than that. It seems like after this CL there should be a cleanup where we make sure our names of our executables are accurate.

Unfortunately this script became kind of a dumping ground for any hacks or emergency fixes to the data that were needed as the benchmark collapse happened, so this is just continuing that tradition. Agreed that we need to clean this up.

Patch-set: 5
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>