Crash GPU process if renderer hangs.
In the Telemetry-based GPU tests, the most common reason that a test
times out is that the renderer process hangs. These hangs are in turn
mostly caused by GPU process hangs. In these situations, a symbolized
minidump is generated for the renderer, which is useful for debugging.
This CL does the same for the GPU process.

It relies on a new DevTools protocol message being added in a
dependent CL.

Tested by manually injecting a GPU process hang in a WebGL conformance
test and verifying that the GPU process's symbolized minidump showed
up in the test output.

Bug: chromium:911772
Change-Id: I3f0773b67c4f00328351872f73c56654ba38cfa3