Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> lgtm for the moment. 
> 
> However, do we really want to support RetryOnFailure along with any other type of result? It seems like we want RetryOnFailure to be mutually exclusive with supporting > 1 kind of expected result; a given test suite will either want the one semantics or the other, right?

Well one use case I can think of is having RetryOnFailure combined with Failure. This so that a test with expectation of [RetryOnFailure Failure] can do all fail all flaky retries of a test and not have it shown as a regression. There seems to be a need for this on some GPU tests which can be seen in this bug, crbug.com/783263.

Patch-set: 1
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>