Changes Typ so that the expectation of a retry on fail test is an eventual Pass
When an expectations file is passed in and a test has the RetryOnFailure
expectation tag, Typ needs to have output showing that it is expecting a
pass. Pass should not be unexpected for a retry on fail test and should not
be a regression.

example:

expectations.txt
# tags: [foo]
crbug.com/123 [foo] test_pass [RetryOnFailure]

Typ output:
[1/1] pass_test.PassingTest.test_pass passed 0.4752s


R=dpranke@chromium.org

Bug: chromium:698902
Change-Id: I45bbf532fe53644bc4c77701c3cfd9174fbae6cb