Update patch set 6
Patch Set 6:

Hi 

This CL changes the expected result of a retry on fail test to be an eventual pass in one of it's flaky retries. Therefore Typ's output should reflect that. For instance it should say the test 'passed' if the test passes and 'failed unexpectedly' if the test fails. Currently GPU tests behave that way. If a test fails all its flaky retries then the GPU test harness throws an exception and Typ see's it as an unexpected failure. If it passes one of the retries it returns from the GPU test harness and Typ says it is a pass.

Best Regards
Rakib Hasan

Patch-set: 6
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false