Changes Typ so that the expectation of a retry on fail test is an eventual Pass
When an expectations file is passed in and a test has the RetryOnFailure
expectation tag, Typ needs to have output showing that it is expecting a
pass. Pass should be expected for a retry on fail test.

example:

expectations.txt
# tags: [foo]
crbug.com/123 [foo] test_pass [RetryOnFailure]

Typ pass output:
...
[1/1] test_pass passed 0.4752s
...

Typ fail output:
...
[1/1] test_pass failed unexpectedly 0.5654s
...

R=dpranke@chromium.org

Bug: chromium:698902
Change-Id: I45bbf532fe53644bc4c77701c3cfd9174fbae6cb