speed: Reduce how often network replay is set up.
Network-replay set up can take 5+ seconds for each page. However, the
replay does not need to be re-setup if the archive for the current story
has not change since the previous story. So to minimize the number of
times network-replay needs to be set-up, sort the stories based on the
archive they are in. This saves ~5-10 minutes for a complete
rendering.desktop run, which is approx. 5-10% of total runtime.

Bug: chromium:#932409
Change-Id: I8a1606fc4656a69002e2cfde38a61346e1451f3a