Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Would it be possible to automatically exclude properties that have been removed from the model to prevent this kind of error in the future?

Think this is all done via the _GetOrCreate which uses ndb.Model.to_dict to exclude items when serializing. Might be able to create a dummy new one or something, and then use the union of the keys in that and the old one to figure out what to strip, maybe something to try in the future?

Patch-set: 1
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>