Update patch set 4
Patch Set 4: Code-Review+1

(3 comments)

In one of your test cases, you have:

    b1.s1 [ RetryOnFailure ]
    b1.s1 [ Failure ]

and you're saying that should be interpreted as:

    b1.s1 [ RetryOnFailure Failure ]

and not

    b1.s1 [ RetryOnFailure Failure Pass ]

I think this is correct, but we probably need to explicitly give an example of this in the spec, and emphasize that the default value of "Pass" is applied only *after* all of the matching is done and if we determine that no values were specified.

Separately, I am reminded that I think kbr@ wants there to be no duplicates allowed in the expectations files for the GPU tests, but we don't actually have that implemented yet (we still need to add this, and this needs to be optional/configurable, since we want to allow duplicates in the layout tests). Once we've implemented that, I think we need to make sure that the above still works, but that

    b1.s1 [ Pass ]
    b1.s1 [ Failure ]

works for the layout tests but not for the GPU tests.

@kbr, @estaab - does this sound right to you? In particular, I'm trying to think ahead to when we also support `[ Slow ]`, which I think we want to work the same way.

Assuming we all agree with the above analysis, I think this patch looks correct.

Patch-set: 4
Reviewer: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1128941 <1128941@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1