Update patch set 2
Patch Set 2: Code-Review-1

Sorry, but I don't think this is the right approach to take here.

Caching things like this has caused problems in the past (see e.g. crbug.com/798499).

The real solution is to *reduce* the number of InstallApplication calls made. Btw, thanks for the trace stacks, those are supper useful!

As you can see the problem is we have three calls to "browser_finder.FindBrowser", when we should only ever try to find the browser once and stick with it.

The existing calls are:
1. In _GetVariationsBrowserArgs, that is related to crbug.com/801513. This call to FindBrowser should be possible to remove by finding an alternative way to get the target_os there.
2. Is story_runner.RunBenchmark; this is the main call we probably want to keep.
3. In shared_page_state.GetPossibleBrowser; we can probably remove this call by making the shared state receive the possible browser that was *already found* by the story_runner in 2.

Apologies, I know this is a bit messy. :(

Patch-set: 2
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1