Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Okay, here's another try. This requires that the generated bin/XXX_bundle script is used as the "apk" rather than the apks/XXX.aab file, but it means there are no script dependencies on chromium. This depends on crrev.com/c/1477022 to add manifest dumping to bundles.
> 
> I like the direction where this is going: using the generated bin script in place of the apk itself. From Telemetry's point of view this would make sense even for regular (non-bundle) apks.
> 
> But I also worry that currently this is getting somewhat cyclical. E.g. apk_operations.py uses devil.Install to install apks, and devil.Install uses apk_operations.py to install bundles.
> 
> John, is that still OK? or any suggestions for a way to get out of this?

It's not quite cyclical in this case, as we go devil.Install(bundle) -> apk_operations -> unpack apks -> devil.Install(apk).

Unpacking the apks requires signature information, which AFAICT is found in the build tree and so very difficult to plumb through devil. I think the issue is that devil assumes the installation of a hermetic thing; an APK is a hermetric thing but an AAB is not while the bundle script is (if one overlooks like recursive use of devil).

Patch-set: 7
Reviewer: Gerrit User 1154589 <1154589@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>