Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Sorry, but I don't think this is the right approach to take here.
> 
> Caching things like this has caused problems in the past (see e.g. crbug.com/798499).
> 
> The real solution is to *reduce* the number of InstallApplication calls made.

I agree. Although this feels like a relatively safe change with fairly explicit code (unlike cache-decorator) ... but also, thousand paper cuts and all that :)

> Btw, thanks for the trace stacks, those are supper useful!
> 
> As you can see the problem is we have three calls to "browser_finder.FindBrowser", when we should only ever try to find the browser once and stick with it.
> 
> The existing calls are:
> 1. In _GetVariationsBrowserArgs, that is related to crbug.com/801513. This call to FindBrowser should be possible to remove by finding an alternative way to get the target_os there.
> 2. Is story_runner.RunBenchmark; this is the main call we probably want to keep.
> 3. In shared_page_state.GetPossibleBrowser; we can probably remove this call by making the shared state receive the possible browser that was *already found* by the story_runner in 2.
> 
> Apologies, I know this is a bit messy. :(

re 1: it looks like that happens in response to the call to CustomizeBrowserOptions in the following code in story_runner.RunBenchmark() [1]:

  331 benchmark.CustomizeBrowserOptions(finder_options.browser_options)
  332 
  333 benchmark_metadata = benchmark.GetMetadata()
  334 possible_browser = browser_finder.FindBrowser(finder_options)

Could the call to CustomizeBrowserOptions() be moved to after the call to FindBrowser(), and possible_browser.target_os be passed in as an argument to CustomizeBrowserOptions()? Looking at the various CustomizeBrowserOptions() overrides, it looks like it is primarily used to include additional command-line flags for the browser, which doesn't really affect the browser-finder(s). So maybe this would be a safe change to make?

[1] https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/internal/story_runner.py?l=331

Patch-set: 2
Reviewer: Gerrit User 1000111 <1000111@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0