speed: Look only for the appropriate devices/browsers.
When running a benchmark, look only for the devices/browsers that
are supported by the benchmark. For example, when running a desktop
benchmark (e.g. rendering.desktop), the code still tries to discover
the android devices attached to the workstation (which takes ~2
seconds), and then the list of browsers appropriate for that device
(which can take ~10 seconds, if you have an android device attached).
This is unnecessary, since these android devices or browsers are not
going to be used anyway. So avoid looking for these devices/browsers.
This brings down the runtime at startup by ~15 seconds. So running a
single story goes down from ~45 seconds to ~30 seconds.

BUG=chromium:932409

Change-Id: I8a48a49bbec8b69d5283d87956cf3d27241b9b4e