speed: Use thread-pool for processing results.
The work-flow currently is:
 1. Run telemetry test for a story.
 2. Collect the trace events.
 3. Process the traces, and produce result.
 4. Collate the result with results from previous stories.
 5. Go back to step 1.

Between steps 5 and step 1, the browser is restarted, and telemetry
waits for the devtools connection to happen etc. So it would be useful
to parallelize steps 3 + 4 with that work. This CL introduces a
thread-pool for asynchronously processing the trace events. Each task
in the thread-pool and running the story (i.e. step 1 above) uses a
shared lock, so that the expensive task of processing the traces do not
happen while the story is being run, to avoid affecting the tests. This
helps the runtime quite a bit, reducing the runtime for benchmarks on
the desktop (e.g. loading.desktop, rendering.desktop) by ~20%.

More runtime data: https://docs.google.com/spreadsheets/d/1Lz3yJ5AoMJ5aT_XraOQtsn5CPB0CgJzuga1sodf6gGY/edit#gid=0

BUG=chromium:932409

Change-Id: I6fa9427c3e6552588fcfd89172d5cf4f86d9f64d