Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> Hi! Here's another proposal that may be a bad idea to speed things up :) PTAL (still a bit wip, but I want to get early feedback before spending too much more time on it).

Thanks! This is a great idea! Unfortunately I think this is not the right time to implement it. :(

As part of [1] we're currently working on splitting the "trace producing" and "trace processing" parts of Telemetry; so the metric computation that this CL is changing to run async will be moved to the end after all stories have been run.

When the split is completed this would allow an even better optimization, by running metric computation for all stories in parallel at once; also without worrying that the cpu used in metric computation might affect other stories still running.

Because this change is not trivial I would rather wait until the splitting is done, and then the better solution can be implemented.

[1]: https://docs.google.com/document/d/11uQA2NVEv84jl84a_kHA0eTx3MRfbxylr83Tr_9UCV4/edit#heading=h.he5yjpxfzuuy

Patch-set: 4
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>