Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Can you add a link in the CL description directly to a file that produces this error?
> I'd like to try to find root cause if possible.

The best I can come up with is:

https://chrome-isolated.appspot.com/browse?namespace=default-gzip&digest=7d87468747424c6fa2c91cd498d2c20b68a961aa&as=perf_results.json

(which is linked in the issue).

It turns out that while we support the `allBins` to be a mapping instead of a list, we don't check whether the bin index actually exists in the bins we know about so far. It could be that we should be adding the indices as we go, instead of directly attempting to index it?

Patch-set: 1
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0