Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > Hi! Here's another proposal that may be a bad idea to speed things up :) PTAL (still a bit wip, but I want to get early feedback before spending too much more time on it).
> 
> Thanks! This is a great idea! Unfortunately I think this is not the right time to implement it. :(
> 
> As part of [1] we're currently working on splitting the "trace producing" and "trace processing" parts of Telemetry; so the metric computation that this CL is changing to run async will be moved to the end after all stories have been run.
> 
> When the split is completed this would allow an even better optimization, by running metric computation for all stories in parallel at once; 

That's pretty interesting! Thank you for sharing! Is there a timeline for when this is expected to happen?

> also without worrying that the cpu used in metric computation might affect other stories still running.
> 

Note that in this CL, the parallel work retains a lock, and the RunStory() also takes the lock, which means (if I understood the python documentation correctly) the metrics computation should not affect the actual benchmark. (an earlier patchset, patchset 2, did not have this lock, which can make the results unreliable).

The runtime for loading.desktop drops from ~52 minutes to ~42 minutes with this change. So I am kind of excited about the reduced cycle times (and hopefully that results in more bot availability). If the long-term fix isn't expected to happen soon-ish, perhaps we can move forward with this?

> Because this change is not trivial I would rather wait until the splitting is done, and then the better solution can be implemented.
> 
> [1]: https://docs.google.com/document/d/11uQA2NVEv84jl84a_kHA0eTx3MRfbxylr83Tr_9UCV4/edit#heading=h.he5yjpxfzuuy

Patch-set: 4
Reviewer: Gerrit User 1000111 <1000111@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0