Add /_ah/warmup
appengine gets /_ah/warmup when it spins up a new instance, both in
prod and dev_appserver. Currently, this always results in 404.
This CL implements a handler for this endpoint.
In both prod and dev, this handler warms some caches like Descriptor's
stored_objects.
In dev, this handler creates some fake data: a Sheriff, a ReportTemplate, a few
TestMetadata, a few Anomalies, and a few hundred Histograms and Rows. If any of
the Anomalies were triaged, they are un-triaged.
The user may also get /_ah/warmup?clear=clear in order to completely reset the
datastore.
This will allow full manual integration testing in dev.
This also demonstrates for clients how to construct valid HistogramSets.

Bug: chromium:918193
Change-Id: I8a87ec2a4aa43a4635f012f3e334025fa542200a