Fix brush handles in spa/chart-base.
chart-base currently uses transform:translateX to position brush handles. This
approach allows updating the position to be faster by skipping layout.
However, this approach requires brush handles to be as wide as the chart.
Translating them to the right causes horizontal scroll bars and large mysterious
empty space to appear.
This CL changes brush handles to use the `left` css property so that they can be
the correct size and avoid creating large empty space and horizontal scroll bars.
This incurs layout when positioning, but the cost is minimal due to
position:absolute.

Bug: chromium:918193
Change-Id: Ifdb0f3ddae3392158a5369e9653cc949426fbad0