HistogramSets: Validate loop index
Before this change, we assumed that the index provided in the histogram
set `allBins` when not a list is always valid. This change introduces a
test with a contrived "bad" example to check that we only use valid
indices. Unfortunately this also means we drop bad data instead of
raising an issue.

R=simonhatch@chromium.org

Bug: chromium:933096
Change-Id: I56e7a7d3d705740df5c5f953457cc9e4c97131dd