Avoid crashing when no crashpad_stackwalker
Avoid throwing an exception from GetStackTrace() when
build/android/stacktrace/ is not there.  This will allow to move parts
of the stackwalker incrementally to either $CHROMIUM_OUTPUT_DIR or to
the necessary isolates. We have not decided which of the two ways to
use, but in any case would like to unblock introducing more tools into
$CHROMIUM_OUTPUT_DIR, as done in:
    https://chromium-review.googlesource.com/c/chromium/src/+/1486353

Also change os.path.exists to os.access(_, os.X_OK) in adjacent places
for consistency with the newly introduced os.access().

Bug: chromium:925453
Change-Id: I921c47878c2c157fbac60b3c1574c9b76ab837cd