speed: Look only for the appropriate devices/browsers.
This CL was landed earlier in http://crrev.com/c/1474597, then reverted
in https://crrev.com/c/1481933, relanded in https://crrev.com/c/1481961
and then reverted again in https://crrev.com/c/1487912

Original CL description:

When running a benchmark, look only for the devices/browsers that
are supported by the benchmark. For example, when running a desktop
benchmark (e.g. rendering.desktop), the code still tries to discover
the android devices attached to the workstation (which takes ~2
seconds), and then the list of browsers appropriate for that device
(which can take ~10 seconds, if you have an android device attached).
This is unnecessary, since these android devices or browsers are not
going to be used anyway. So avoid looking for these devices/browsers.
This brings down the runtime at startup by ~15 seconds. So running a
single story goes down from ~45 seconds to ~30 seconds.

BUG=chromium:932409

Change-Id: I32aaff090d944e4085d93a2bc90040ed759756b5