Change the number of test inputs sent to the process pool at one time
Currently the number of test inputs sent to the process pool at one time
is equal to value passed through the --jobs command line argument. The
default value for this command line argument is set to the number CPUs
on the machine. This means that serially executed browser tests will not
run serially if the value for the argument is not set to one. This could
cause tests to time out because resources could become unavailabe due
to the load of tests running in parallel.

Change-Id: I469f6b697f3df083ec24a15678245510cefbcc22
Bug: chromium::935814