Fixes issue with serially executed browser tests
Currently the number of jobs created in a process pool and number of
jobs sent to the pool at a time is equal to the value passed for the
--job command line argument. The default value for this argument is
set to the number of CPUs on the machine. This means that serially
executed jobs are not actually running serially if the value of the
command line argument is not equal to one. This can cause tests to
fail or timeout becauses resources may become unaavaialbe due to the
load of tests running.

Bug: chromium:935814
Change-Id: I469f6b697f3df083ec24a15678245510cefbcc22