Update patch set 4
Patch Set 4:

(1 comment)

I think when we had talked about this before, we talked about the prefix needing to apply to names generally, not just in the expectations files. 

In other words, we need to make sure the names that are output, stored in the json results, and accepted as input in the filter files and on the command line all use the same variant of the names. I don't want us using the long version in some places and the shorter version elsewhere.

Given that, "--append-prefix-to-all_test_expectation_patterns" is really long and ends up not really being that accurate. Can we use something like --prefix or --test-prefix or --name-prefix or --namespace instead?

Patch-set: 4
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>