Add --dont-root-device command line argument to Telemetry.
This avoids doing all of the operations which require root access. It
switches the command line file to the version in /data/local/tmp, and
also places the user profile in a directory under /data/local/tmp.

Tested locally by running a subset of the WebGL conformance tests on a
non-rooted Pixel 2 with Chrome Canary:

  ./content/test/gpu/run_gpu_integration_test.py webgl_conformance \
    --browser=android-chrome-canary --test-filter=conformance_attribs \
    --dont-root-device

Bug: chromium:826295
Change-Id: Ifdfbda7697f6336e1732f60cff8f078f276ad6bf