Update patch set 3
Patch Set 3: Commit-Queue+1

(3 comments)

> Patch Set 2:
> 
> (3 comments)
> 
> Please write a design doc for this and get it approved by myself and by jbudorick@ (or some other devil dev) and by perezju@. We need to think about whether this is the right longterm plan.

I'll be happy to put together a design doc tomorrow.

> Also, it would be nicer to have fewer code paths here. For example, why don't we just always use the "legacy path"? If there isn't a good reason for that, then we should just switch. A design doc will help us figure out those things.

The "legacy path" is the one we want to get rid of. One question is whether the oldest versions of Chrome that Telemetry is still supposed to work with support the new location of the command line flags in /data/local/tmp/chrome-command-line, as opposed to the old location of /data/local/chrome-command-line, which requires root access on the device in order to write.

I was nervous about changing the default behavior of Telemetry on the bots, as there are a large number of users both inside and outside of the Chromium project.

This CL allows our group to run GPU tests on phones which we can't get root access to (like off-the-shelf consumer phones), which has been a longstanding feature request. It has been tested locally.

Patch-set: 3
CC: Gerrit User 1001921 <1001921@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1