Fork groupAlerts() for spa
The V1 /alerts page uses d.groupAlerts().
V2spa uses a different structure for alerts with a Descriptor, and needs
to pass a flag into groupAlerts() when displaying triaged alerts.
This CL forks groupAlerts to support the new structure and the new flag.
The old groupAlerts() can be removed when V1 is deprecated. The grouping
algorithm doesn't change often, so version skew bugs are unlikely.

Bug: chromium:918193
Change-Id: I618ca28e53a80f6ba170c0da91170b8923baa2b4