Adds new command line argument to Typ
Adds the --test-name-prefix command line argument to Typ. This
command line argument is used to pass a prefix string that can
be removed from the front of a test's name. In this CL, the string
passed as the argument will be removed from the test name in the
output that happens when test is finished.

R=dpranke@chromium.org

Bug: chromium:698902
Change-Id: I12f053942ec2e4d7a77be7742da6f71edc2e8fac