Update patch set 6
Patch Set 6:

Hi Dirk

This the first CL of a series that will add the --test name-prefix and its functionality. In this CL I remove the prefix from the test name in the output that happens when a test is finished.

Regards
Rakib Hasan

Patch-set: 6
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false