Update patch set 6
Patch Set 6:

Hi
This is the first of series of CL's that adds the --test-name-prefix command line argument to Typ. In this CL, the the prefix passed as the argument value is removed from the test name in the output that happens when a test is finished.

Regards
Rakib Hasan

Patch-set: 6
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>