Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > Not to disagree with Dirk but I think this is going to break FindIt. If --test-filter doesn't accept the short names then there's a basic disagreement about the test's name.
> 
> I'm not sure I understand the purpose of this flag.  Does the flag shorten the test names in test results?
> 
> For Findit it basically gets test names from test results and triggers a rerun task on  some of the tests using --test-filter. 
> 
> If Findit gets shortened test names from test results, and these names cannot be recognized by test runners, yes Findit will be broken.
> 
> And on a side not, we actually have an opposite case right now: for telemetry benchmarks, Findit gets some long version of test names but test runner only recognizes a shorter version of name. (See https://bugs.chromium.org/p/chromium/issues/detail?id=894254#c41 point 2)

Currently when Typ outputs to stdout that a test is finished. It includes the fully qualified 
name for a test. 

For example 
[1/10] gpu_tests.webgl_conformance_integration_test.WebGLConformanceIntegrationTest.WebglExtension_EXT_color_buffer_float Passed ...

This Cl will add the --test-name-prefix command line argument which will be used to store the prefix of the test's name. This CL will also remove the test name's prefix from the output Typ's output.

So the above test will become the following if we pass the following command line option 
--test-name-prefix=gpu_tests.webgl_conformance_integration_test.WebGLConformanceIntegrationTest.

[1/10] WebglExtension_EXT_color_buffer_float Passed ...

The whole purpose of adding this command line option was so that we can shorten the test names for expectations files for GPU tests. Repeatedly having  "gpu_tests.webgl_conformance_integration_test.WebGLConformanceIntegrationTest." for about 700 webgl conformance integration test expectations can make the file very verbose and hard to read. However Dirk raised concerns that having different tools looking at different names for tests would cause all the tools to be out of sync. So this is the first CL in a series to that will make sure that when --test-name-prefix command line arg is used, it will shorten the test name everywhere.

I also noticed that in telemetry browser tests --test-filter uses the test method name to filter for tests which can be seen in the link below.

https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/testing/run_browser_tests.py?rcl=3c7b056c0c8bee3b6649194c55d083be1cb9c3ba&l=171

t.shortName() returns the test method name only.

So if Findit can only get the test method name from the results then it should be able to restart the test.

Patch-set: 6